+++ milo/milo.h
@@ -1,9 +1,0 @@
-#ifdef __cplusplus
-extern "C" {
-#endif
-
-char *dtoa_milo(double val);
-
-#ifdef __cplusplus
-}
-#endif